requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
bson